Allow default value of primitive properties to be null
An annotation does not have to declare a default value and, in the
reflection API, the absence of a default definition means the value is
returned as null, even if the member is of a primitive type.

When creating annotation maps from annotation classes, it is necessary
to allow for the possibility that the default (and, consequently, the
value) is null. This has been accommodated by introducing the notion of
an "UNDEFINED" member value.